Relocate GetCompanySummarySuccessPresenter
