+++ arbeitszeit_web/presenters/get_company_summary_presenter.py
